Add stats to routes

+++ test/test-sockets.js
@@ -40,16 +40,10 @@ describe('socket.io server', () => {
   });
 
   after((done) => {
-    try {
-      User.collection.drop();
-    } catch (e) {
-      if (e.code === 26) {
-        console.log('namespace %s not found', User.collection.name);
-      } else {
-        console.log("ERROR: Failed to drop collection!");
-        throw e;
-      }
-    }
+    User.countDocuments({}, (numOfDocuments) => {
+        if (numOfDocuments)
+            User.collection.drop();
+    });
     done();
   });
 
@@ -145,5 +139,4 @@ describe('socket.io server', () => {
       clientB.emit('mac', {macAddress: "FF:FF:FF:FF:FF"});
     }, 3000);
   });
-
 });
